+++ vite.config.ts
@@ -1,5 +1,4 @@
-// vite.config.ts
-// vite.config.ts 2.0.3
+// vite.config.ts 2.0.4
 
 // This Vitest configuration is designed for a TypeScript project.
 
@@ -8,8 +7,21 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   cacheDir: '.cache/vitest',
   test: {
-    include: ['**/vitest/**/*.spec.{ts,mts,cts}', '**/vitest/**/*.test.{ts,mts,cts}'],
-    exclude: ['**/.cache', '**/apps/', '**/build', '**/chip', '**/coverage', '**/dist/', '**/node_modules/', '**/screenshots', '**/temp', '**/vendor'],
+    include: ['**/vitest/**/*.{spec,test}.{ts,mts,cts}'],
+    exclude: [
+      '**/.cache/',
+      '**/apps/',
+      '**/build/',
+      '**/chip/',
+      '**/coverage/',
+      '**/dist/',
+      '**/node_modules/',
+      '**/screenshots/',
+      '**/scripts/',
+      '**/src/mock/',
+      '**/temp/',
+      '**/vendor/',
+    ],
     globals: true,
     clearMocks: true,
     restoreMocks: true,
@@ -19,7 +31,21 @@ export default defineConfig({
       provider: 'v8',
       reporter: ['text', 'lcov'],
       include: ['**/src/**/*.{ts,mts,cts}'],
-      exclude: ['**/.cache', '**/apps/', '**/build', '**/chip', '**/coverage', '**/dist/', '**/node_modules/', '**/screenshots', '**/temp', '**/vendor', '**/src/**/*.d.ts'],
+      exclude: [
+        '**/.cache/',
+        '**/apps/',
+        '**/build/',
+        '**/chip/',
+        '**/coverage/',
+        '**/dist/',
+        '**/node_modules/',
+        '**/screenshots/',
+        '**/scripts/',
+        '**/src/mock/',
+        '**/temp/',
+        '**/vendor/',
+        '**/src/**/*.d.ts',
+      ],
       thresholds: {
         statements: 100,
         branches: 100,